Update and rename csv_parser.js to parser.js
#471 comments
